refactor(egress): consolidate egress UI onto /egress with optional env filter
Move all egress controls (firewall toggle, policy cards, rules, traffic feed)
off the per-environment Egress tab and onto the top-level /egress page.
The page shows every environment's egress at once and lets the user filter
to a single environment via a `?env=<id>` URL param. The environment detail
page no longer has an Egress tab.

Splits egress-tab.tsx into three reusable components
(egress-firewall-card, egress-policy-card, egress-traffic-feed) and drops
the now-unused tab and its smoke tests. Backend is unchanged — both
/api/egress/policies and /api/egress/events already accept an optional
environmentId query param.
