+++ client/src/app/environments/[id]/page.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useParams, Link, Navigate, useSearchParams } from "react-router-dom";
+import { useParams, Link, Navigate } from "react-router-dom";
 import { Environment } from "@mini-infra/types";
 import { useEnvironment } from "@/hooks/use-environments";
 import { StacksList } from "@/components/environments";
@@ -7,7 +7,6 @@ import { useUserStacks } from "@/hooks/use-applications";
 import { EnvironmentEditDialog } from "@/components/environments/environment-edit-dialog";
 import { EnvironmentDeleteDialog } from "@/components/environments/environment-delete-dialog";
 import { RemediateHAProxyDialog } from "@/components/haproxy/remediate-haproxy-dialog";
-import { EgressTab } from "@/components/egress/egress-tab";
 import {
   Card,
   CardContent,
@@ -18,7 +17,6 @@ import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Skeleton } from "@/components/ui/skeleton";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -38,15 +36,12 @@ import {
   IconCloud,
   IconWorldWww,
   IconHome,
-  IconShield,
 } from "@tabler/icons-react";
 import { useFormattedDate } from "@/hooks/use-formatted-date";
 import { cn } from "@/lib/utils";
 
 export function EnvironmentDetailPage() {
   const { id: environmentId } = useParams<{ id: string }>();
-  const [searchParams] = useSearchParams();
-  const initialTab = searchParams.get("tab") === "egress" ? "egress" : "overview";
 
   const [editDialogOpen, setEditDialogOpen] = useState(false);
   const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
@@ -73,12 +68,6 @@ export function EnvironmentDetailPage() {
   const userStacks = (userStacksData?.data ?? []).filter(
     (s) => s.environmentId === environmentId,
   );
-
-  // Gate the Egress tab behind egress:read.
-  // Browser sessions have full access (null permissions = full access).
-  // When API key auth surfaces session permissions to the client in a future
-  // slice, replace `true` with: hasPermission(session.permissions, "egress:read")
-  const canReadEgress = true;
 
   if (!environmentId) {
     return <Navigate to="/environments" replace />;
@@ -325,27 +314,7 @@ export function EnvironmentDetailPage() {
       </div>
 
       <div className="px-4 lg:px-6 max-w-full">
-        <Tabs defaultValue={initialTab}>
-          <TabsList className="mb-4">
-            <TabsTrigger value="overview">Overview</TabsTrigger>
-            {canReadEgress && (
-              <TabsTrigger value="egress" className="flex items-center gap-1.5">
-                <IconShield className="h-3.5 w-3.5" />
-                Egress
-              </TabsTrigger>
-            )}
-          </TabsList>
-
-          <TabsContent value="overview">
-            <StacksList environmentId={environment.id} />
-          </TabsContent>
-
-          {canReadEgress && (
-            <TabsContent value="egress">
-              <EgressTab environmentId={environment.id} />
-            </TabsContent>
-          )}
-        </Tabs>
+        <StacksList environmentId={environment.id} />
       </div>
 
       {/* Dialogs */}
