+++ client/src/user-docs/ui-elements/manifest.json
@@ -452,7 +452,7 @@
       "id": "environment-egress-firewall-toggle",
       "label": "Environment Egress Firewall Toggle",
       "elementType": "Switch",
-      "file": "client/src/components/egress/egress-tab.tsx"
+      "file": "client/src/components/egress/egress-firewall-card.tsx"
     },
     {
       "id": "header-backup-health",
